move to dungeon_items.c
